Change bytes to uints
The config options were originally meant to be positive only numbers never going above the range of an byte, but this should avoid limitations due to being able to set the tick rate to values outside of a byte's range.

+++ DynamicTickRate/Config.cs
@@ -12,29 +12,29 @@ public partial class DynamicTickRate : ResoniteMod
                 () => true);
 
     [AutoRegisterConfigKey]
-    public static readonly ModConfigurationKey<Byte> MinTickRate =
-            new ModConfigurationKey<Byte>(
+    public static readonly ModConfigurationKey<uint> MinTickRate =
+            new ModConfigurationKey<uint>(
                 "MinTickRate",
                 "The lowest the tick rate is allowed to be set by this mod",
                 () => 30);
 
     [AutoRegisterConfigKey]
-    public static readonly ModConfigurationKey<Byte> MaxTickRate =
-            new ModConfigurationKey<Byte>(
+    public static readonly ModConfigurationKey<uint> MaxTickRate =
+            new ModConfigurationKey<uint>(
                 "MaxTickRate",
                 "The highest the tick rate is allowed to be set by this mod",
                 () => 60);
 
     [AutoRegisterConfigKey]
-    public static readonly ModConfigurationKey<Byte> AddedTicksPerUser =
-            new ModConfigurationKey<Byte>(
+    public static readonly ModConfigurationKey<uint> AddedTicksPerUser =
+            new ModConfigurationKey<uint>(
                 "AddedTicksPerUser",
                 "How much the tick rate is increased per user",
                 () => 2);
 
     [AutoRegisterConfigKey]
-    public static readonly ModConfigurationKey<Byte> AddedTicksPerWorld =
-            new ModConfigurationKey<Byte>(
+    public static readonly ModConfigurationKey<uint> AddedTicksPerWorld =
+            new ModConfigurationKey<uint>(
                 "AddedTicksPerWorld",
                 "How much the tick rate is increased per world",
                 () => 4);
